SteamNexus/requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
networkx
nltk
kaggle
python-dotenv
python-pptx
